tbNewToolbox: disable buttons after tb creation
Disable buttons once the toolbox is created. These buttons (edit / get planned actions / create) areuseless after creation and the disabling gives an appropriate feedback to the user that the action is done. Don't close the app since the text field with the exectued actions might still contain
useful info for the user